cdk-nested-stack/requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws_cloudformation